requirements.txt
reportlab==4.4.1
Pillow==11.2.1